minor changes in biological optimization procedure

+++ matRadGUI.m
@@ -1273,7 +1273,7 @@ for i = 1:size(OldCst,1)
                     if isempty(data{j,6})
                        FlagValidParameters=false;
                     else
-                         NewCst{Cnt,4}(CntObjF,1).parameter(1,2) = data{j,6};
+                         NewCst{Cnt,4}(CntObjF,1).parameter(1,2) = double(data{j,6});
                     end
                 end
             end
